More changes from pyBEscanner-v3
Internal Bans Storage / Report Bans File
Changed Logs / Temp File location to
<pyBEscanner> directory...
Info @ changes.txt

+++ pyBEscanner.py
@@ -24,7 +24,7 @@ import copy
 import re
 import sys
 
-from modules import logs_battleye, rcon_cscript
+from modules import bans, logs_battleye, rcon_cscript
 
 
 
@@ -65,16 +65,16 @@ class Main:
 		else:
 			open(self.lockfile, 'w').close()
 
-		self.config = ConfigParser.ConfigParser()
-		self.config.read(self.conf_file)
-		if self.config.has_option("Default", "Version"):
-			if self.config.get("Default", "Version") != "13":
+		config = ConfigParser.ConfigParser()
+		config.read(self.conf_file)
+		if config.has_option("Default", "Version"):
+			if config.get("Default", "Version") != "15":
 				print "-------------------------------------------------"
 				print "ERROR: Bad conf/servers.ini version"
 				print "-------------------------------------------------"
 				print "Read Changes.txt for more info"
-				print "Old version = " + self.config.get("Default", "Version")
-				exit()
+				print "Old version = " + config.get("Default", "Version")
+				sys.exit()
 		else:
 			print "-------------------------------------------------"
 			print "ERROR: No servers.ini version found"
@@ -84,15 +84,22 @@ class Main:
 			print
 			print "Or if u haven't updated in awhile"
 			print "Recommend u delete pyBEscanner temp folders & read Changes.txt for update changes"
-			exit()
+			sys.exit()
 
 	def loadconfig(self):
 
-		self.config.read(self.conf_file)
-		self.server_settings = self.config.sections()
-		self.server_settings.remove("Default")
+		config = ConfigParser.ConfigParser()
+		config.read(self.conf_file)
+		config_sections= config.sections()
+		config_sections.remove("Default")
 
-		default = {}
+		default = {"pyBEscanner Directory": self.main_dir}
+		self.server_settings = []
+		# Ban.txt  File Location
+		if config.has_option("Default", "Bans Directory"):
+			default["Bans Directory"] = config.get("Default", "Bans Directory")
+			default["Bans"] = None
+			# Doesnt initializes bans, until encounters a server with symlink setting turned on
 
 		options = [["Scan Addbackpackcargo", "addbackpackcargo"],
 					["Scan Addmagazinecargo", "addmagazinecargo"],
@@ -113,55 +120,66 @@ class Main:
 					["OffSet", "OffSet"],
 					["Ban Message", "Ban Message"],
 					["Kick Message", "Kick Message"],
+					["Report Message", "Report Message"],
 					["Ban IP", "Ban IP"],
-					["Filters", "Filters"]]
+					["Filters", "Filters"],
+					["Bans Symlinked", "Bans Symlinked"]]
 
 		## Scan Settings -- Default
-		self.interval = int(self.config.get("Default", "interval", "60"))
+		self.interval = int(config.get("Default", "interval", "60"))
 
 		for x in range(len(options)):
-			default[options[x][1]] = self.config.get("Default", options[x][0])
+			default[options[x][1]] = config.get("Default", options[x][0])
 
 		## Debug Settings
-		if self.config.has_option("Default", "Debug File"):
-			self.debug_file = self.config.get("Default", "Debug File")
+		if config.has_option("Default", "Debug File"):
+			self.debug_file = config.get("Default", "Debug File")
 		else:
 			self.debug_file = os.path.join(self.main_dir, "debug.log")
 
-		if self.config.has_option("Default", "Debug Level"):
-			self.debug_level = self.config.get("Default", "Debug Level")
+		if config.has_option("Default", "Debug Level"):
+			self.debug_level = config.get("Default", "Debug Level")
 		else:
 			self.debug_level = "WARNING"
 
-		x = 0
-		while x < (len(self.server_settings)):
-			temp = copy.copy(default)
+		for section in config_sections:
+			server = copy.copy(default)
 
 			## Server Info
-			temp["ServerName"] = self.config.get(self.server_settings[x], "ServerName")
-			temp["ServerIP"] = self.config.get(self.server_settings[x], "ServerIP")
-			temp["ServerPort"] = self.config.get(self.server_settings[x], "ServerPort")
-			temp["RconPassword"] = self.config.get(self.server_settings[x], "RconPassword")
-			temp["BattlEye Directory"] = self.config.get(self.server_settings[x], "BattlEye Directory")
-			temp["Temp Directory"] = os.path.join(temp["BattlEye Directory"], "pyBEscanner", "Temp")
+			server["ServerName"] = config.get(section, "ServerName")
+			server["ServerIP"] = config.get(section, "ServerIP")
+			server["ServerPort"] = config.get(section, "ServerPort")
+			server["RconPassword"] = config.get(section, "RconPassword")
+			server["BattlEye Directory"] = config.get(section, "BattlEye Directory")
+			server["Server Console Log"] = config.get(section, "Server Console Log")
+			server["Server RPT Log"] = config.get(section, "Server RPT Log")
+			server["Temp Directory"] = os.path.join(self.main_dir, "temp", section)
+			server["Logs Directory"] = os.path.join(self.main_dir, "logs", section)
+			server["LockFile"] = os.path.join(server["Temp Directory"], "scan-stopped.lockfile")
+			server["LockFile-Ask"] = os.path.join(server["Temp Directory"], "scan-stop-ask.lockfile")
 
 			for y in range(len(options)):
-				if self.config.has_option(self.server_settings[x], options[y][0]):
-					temp[options[y][1]] = self.config.get(self.server_settings[x], options[y][0])
+				if config.has_option(section, options[y][0]):
+					server[options[y][1]] = config.get(section, options[y][0])
 
-			temp["Filters"] = re.sub(",\s*", ",", temp["Filters"])
-			temp_filters = temp["Filters"].split(",")
-			temp["Filters"] = []
+			server["Filters"] = re.sub(",\s*", ",", server["Filters"])
+			temp_filters = server["Filters"].split(",")
+			server["Filters"] = []
 			for filters in temp_filters:
 				if filters == "Custom":
-					temp["Filters"].append(os.path.join(temp["BattlEye Directory"], "pyBEscanner", "filters", "Custom"))
+					server["Filters"].append(os.path.join(server["BattlEye Directory"], "pyBEscanner", "filters"))
 				else:
-					temp["Filters"].append(os.path.join(self.main_dir, "filters", filters))
+					server["Filters"].append(os.path.join(self.main_dir, "filters", filters))
 
-			self.server_settings[x] = temp
+			if server["Bans Symlinked"] == "on":
+				if default["Bans"] == None:
+					default["Bans"] = bans.Bans(default["Bans Directory"])
+				server["Bans"] = default["Bans"]
+			else:
+				default["Bans"] = bans.Bans(server["BattlEye Directory"])
 
 			# Generated Settings
-			self.server_settings[x]["Battleye Logs"] = ["addbackpackcargo",
+			server["Battleye Logs"] = ["addbackpackcargo",
 											"addmagazinecargo",
 											"addweaponcargo",
 											"attachto",
@@ -178,32 +196,32 @@ class Main:
 											"setvariable",
 											"teamswitch"]
 
-			self.server_settings[x]["Battleye Logs Location"] = {}
-			self.server_settings[x]["Battleye Temp Logs"] = {}
-			self.server_settings[x]["Battleye Backup Logs"] = {}  # TODO
+			server["Battleye Logs Location"] = {}
+			server["Battleye Temp Logs"] = {}
+			server["Battleye Backup Logs"] = {}  # TODO
 
-			self.server_settings[x]["Banlist Filters"] = {}
-			self.server_settings[x]["Kicklist Filters"] = {}
-			self.server_settings[x]["Whitelist Filters"] = {}
-			self.server_settings[x]["Spamlist Filters"] = {}
+			server["Banlist Filters"] = {}
+			server["Kicklist Filters"] = {}
+			server["Whitelist Filters"] = {}
+			server["Spamlist Filters"] = {}
 
 
-			for be_log in self.server_settings[x]["Battleye Logs"]:
-				self.server_settings[x]["Battleye Logs Location"][be_log] = os.path.join(self.server_settings[x]["BattlEye Directory"], (be_log + ".log"))
-				self.server_settings[x]["Battleye Temp Logs"][be_log] = os.path.join(self.server_settings[x]["Temp Directory"], (be_log + ".log"))
+			for be_log in server["Battleye Logs"]:
+				server["Battleye Logs Location"][be_log] = os.path.join(server["BattlEye Directory"], (be_log + ".log"))
+				server["Battleye Temp Logs"][be_log] = os.path.join(server["Temp Directory"], (be_log + ".log"))
 
-				self.server_settings[x]["Banlist Filters"][be_log] = []
-				self.server_settings[x]["Kicklist Filters"][be_log] = []
-				self.server_settings[x]["Whitelist Filters"][be_log] = []
-				self.server_settings[x]["Spamlist Filters"][be_log] = []
+				server["Banlist Filters"][be_log] = []
+				server["Kicklist Filters"][be_log] = []
+				server["Whitelist Filters"][be_log] = []
+				server["Spamlist Filters"][be_log] = []
 
-				for filters in self.server_settings[x]["Filters"]:
-					self.server_settings[x]["Banlist Filters"][be_log].append(os.path.join(filters, be_log + ".banlist"))
-					self.server_settings[x]["Kicklist Filters"][be_log].append(os.path.join(filters, be_log + ".kicklist"))
-					self.server_settings[x]["Whitelist Filters"][be_log].append(os.path.join(filters, be_log + ".whitelist"))
-					self.server_settings[x]["Spamlist Filters"][be_log].append(os.path.join(filters, be_log + ".spamlist"))
+				for filters in server["Filters"]:
+					server["Banlist Filters"][be_log].append(os.path.join(filters, be_log + ".banlist"))
+					server["Kicklist Filters"][be_log].append(os.path.join(filters, be_log + ".kicklist"))
+					server["Whitelist Filters"][be_log].append(os.path.join(filters, be_log + ".whitelist"))
+					server["Spamlist Filters"][be_log].append(os.path.join(filters, be_log + ".spamlist"))
+			self.server_settings.append(server)
 
-			x = x + 1
 
 	def start(self):
 		old_config_timestamp = None
@@ -230,12 +248,26 @@ class Main:
 					scan_count = scan_count + 1
 
 				for server in self.server_settings:
-					bans_file = os.path.join(server["BattlEye Directory"], "bans.txt")
-					if os.path.isfile(bans_file) is False:
-						open(bans_file, 'w').close()
-					server["Bans.txt Timestamp"] = os.path.getmtime(bans_file)
-					server_scan = logs_battleye.Scanner(server)
-					server_scan.scan()
+					if os.path.isfile(server["LockFile"]):
+						print
+						print("LockFile Detected - Skipping " + server["ServerName"])
+						scan_count = 60
+					else:
+						bans_file = os.path.join(server["BattlEye Directory"], "bans.txt")
+						if os.path.isfile(bans_file) is False:
+							open(bans_file, 'w').close()
+						server["Bans.txt Timestamp"] = os.path.getmtime(bans_file)
+						logs_battleye.Scanner(server).scan()
+						if os.path.isfile(server["LockFile-Ask"]):
+							print()
+							print("LockFile Detected, Finished Scanning Logs for Server " + server["ServerName"])
+							#logs_server.ConsoleScanner(server).scan_log(0)
+							#logs_server.RPTScanner(server).scan_log(0)
+							open(server["LockFile"], 'w').close()
+						else:
+							pass
+							#logs_server.ConsoleScanner(server).scan_log()
+							#logs_server.RPTScanner(server).scan_log()
 
 				for server in self.server_settings:
 					kicks_file = os.path.join(server["BattlEye Directory"], "kicks.txt")
@@ -245,14 +277,18 @@ class Main:
 						os.remove(kicks_file)
 						scan_count = 60
 
-				for server in self.server_settings:
-					bans_file = os.path.join(server["BattlEye Directory"], "bans.txt")
-					if os.path.isfile(bans_file) is False:
-						open(bans_file, 'w').close()
-					if server["Bans.txt Timestamp"] != os.path.getmtime(bans_file):
-						rcon = rcon_cscript.Rcon(self.server_settings[x]["ServerIP"], server["ServerPort"], server["RconPassword"])
+				for server in list(self.server_settings):
+					server["Bans"].checkBans()
+					if server["Bans"].getStatus() == True:
+						print
+						print ("Reloading Bans: " + server["ServerName"])
+						logging.info("Reloading Bans")
+						server["Bans"].writeBans()
+						rcon = rcon_cscript.Rcon(server["ServerIP"], server["ServerPort"], server["RconPassword"])
 						rcon.reloadbans()
 						scan_count = 60
+				for server in list(self.server_settings):
+					server["Bans"].updateStatus(False)
 
 				time.sleep(self.interval)
 			except KeyboardInterrupt:
